dashboard/pyproject.toml
[tool.poetry]
name = "sensing-garden-dashboard"
version = "0.1.0"
description = "Dashboard for Sensing Garden Backend"
authors = ["Sensing Garden Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.0.1"
boto3 = "^1.18.0"
python-dotenv = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
